feat(jsonschema): add default id to bundle
When bundling a schema that does not have an explicit identifier,
and a default identifier is provided to the  function,
this change will now make sure that the identifier is added to the
bundled schema.

This makes it easier for consumers of the bundled schema to know
what the identifier of the schema is without having to be aware
of the implicit identifier that was passed to the
function.

Co-Authored-By: Juan Cruz Viotti <jv@jviotti.com>